+++ CHANGELOG.md
@@ -41,6 +41,7 @@ CHANGELOG for Sulu
     * BUGFIX      #1642 [AdminBundle]     Fixed cropping issue in table-view of datagrid
     * BUGFIX      #1641 [MediaBundle]     Fixed types in media-selection
     * ENHANCEMENT #1638 [ContentBundle]   Removed unnecessary variable in content form
+    * FEATURE     #1643 [ContactBundle]   Added title of address
     * FEATURE     #1584 [ContentBundle]   Show UI elements in content management based on security
     * BUGFIX      #1633 [Rest]            Listbuilder: fix for concat joins in where clause;
                                           Fix in sort (order by id as default)